working CRUD functionality

+++ backend/controllers/reviewController.js
@@ -45,10 +45,9 @@ reviewController.postReviews = async (req, res, next) => {
     const queryText =
       "INSERT INTO reviews (landlord_id, text, rating, would_rent_again, date, user_id) values($1,$2,$3,$4,$5,$6) RETURNING landlord_id, text, rating, would_rent_again, date, user_id";
     const value = [landlord_id, text, rating, would_rent_again, date, userID];
-    const review = await db.query(queryText, value).rows[0];
-
-    res.locals.reviews = review;
-    console.log(review);
+    const review = await db.query(queryText, value);
+    res.locals.reviews = review.rows;
+    console.log(res.locals.reviews);
     return next();
   } catch (err) {
     next({
@@ -59,12 +58,9 @@ reviewController.postReviews = async (req, res, next) => {
   }
 };
 
-//  'SELECT AVG(rating) FROM reviews where landlord_id = $1;';
-
 //make another getReview function specific to user
 reviewController.getUserReviews = (req, res, next) => {
   const user = req.params.id;
-  console.log(user);
   const text =
     //'SELECT reviews.* FROM reviews INNER JOIN users ON reviews.user_id = users._id AND reviews.user_id = $1'; working PERFECTLY FINE
     "SELECT reviews.*, landlords.* FROM reviews INNER JOIN users ON reviews.user_id = users._id AND reviews.user_id = $1 INNER JOIN landlords ON reviews.landlord_id = landlords._id";
@@ -107,21 +103,20 @@ reviewController.getReviews = (req, res, next) => {
 
 reviewController.updateReview = async (req, res, next) => {
   try {
-    const { landlord_id, user_id, rating, text, would_rent_again } = req.body;
+    const { landlord_id, userID, rating, text, would_rent_again, date } =
+      req.body;
     const queryText =
       "UPDATE reviews \
-      SET rating = $3, text = $4, would_rent_again = $5 \
+      SET rating = $3, text = $4, would_rent_again = $5, date = $6 \
       WHERE landlord_id = $1 AND user_id = $2 RETURNING *;";
 
-    const value = [landlord_id, user_id, rating, text, would_rent_again];
+    const value = [landlord_id, userID, rating, text, would_rent_again, date];
 
-    console.log(value);
+    const review = await db.query(queryText, value);
 
-    const review = await db.query(queryText, value).rows[0];
+    console.log(review.rows);
 
-    console.log(review);
-
-    res.locals.review = review;
+    res.locals.review = review.rows;
     return next();
   } catch (err) {
     next({
@@ -134,19 +129,17 @@ reviewController.updateReview = async (req, res, next) => {
 
 reviewController.deleteReview = async (req, res, next) => {
   try {
-    const { landlord_id, user_id } = req.body;
-
     const queryText =
-      "DELETE FROM reviews \
-      WHERE landlord_id = $1 AND user_id = $2;";
+      // "DELETE FROM reviews \
+      // WHERE landlord_id = $1 AND user_id = $2;";
+      "DELETE FROM reviews WHERE user_id = $1 AND text = $2";
 
-    const value = [landlord_id, user_id];
+    // const value = [landlord_id, user_id];
+    const value = [req.params.id, req.body.review];
 
-    const review = await db.query(queryText, value).rows[0];
-
-    console.log(review);
-
-    res.locals.user_id = user_id;
+    const deletedReview = await db.query(queryText, value);
+    console.log(deletedReview);
+    res.locals.deletedReview = req.body.review;
 
     return next();
   } catch (err) {
